Context: Made scribe_attach() succeesful in all cases
This allow a better integration in copy_process() because
attach_process() needs the vpid to be valid, which is done by
attach_pid(p, PIDTYPE_PID, pid), and this happens at the very end of
copy_process() where it's difficult to fail.

Also the queues got a "persistent" reference function to make things
cleaner and easier to understand.