Dev: The read/write on the log file are performed by a kthread
We need a kthread for performance reasons: consider a single process
getting recorded. When that process queues some events in its queue, we
want it to immediately return to work, and the actual serialization will
happen on another CPU.

the device's dev_read() and dev_write() are now used as a control
channel. The ioctl() are not used anymore.