Data: x86-32 hooks on userspace data transferts function with recording support
Currently, if event memory allocation fails during the data transfert,
we miserably fail, and abort the whole recording with
scribe_emergency_stop().

All user accesses can be probed by hooking on:
    get_user
    put_user
    __get_user_nocheck
    __put_user_nocheck
    __get_user_size_ex
    __put_user_size_ex
    __copy_to_user
    __copy_from_user

    __copy_to_user_inatomic
    __copy_from_user_inatomic
    __copy_from_user_nocache
    __copy_from_user_inatomic_nocache

    __do_strncpy_from_user
    __do_clear_user
    strnlen_user

We won't touch the inatomic/nocache versions because we don't want to be in
an atomic region (need to allocate memory). Instead we will change the code
that uses inatomic versions to use their slow path.