Resource: filp are resource tracked, read/write access are distinguished
For pipes, it's quite convenient to have two different resources for each
endpoint:
- Reduced overhead (decoupling the readers/writers)
- Removed deadlocking issues when the writer wants to send data to the
  pipe while the pipe buffer is full (the reader will not be able to
  consume the data since the writer has the resource lock).

To be able to lock the right inode resource, read and write version of
locks are now provided.

Because of this multiple resource inode feature, we use the file
pointer as a resource (which is also more natural because we can really
split dependencies between the file pointer and the underlying inode).